+++ abci/app/v1/check_tx.go
@@ -748,6 +748,10 @@ func (app *ABCIApplication) callCheckTx(name string, param []byte, nodeID string
 		return app.reduceNodeTokenCheckTx(param, nodeID)
 	case "SetNodeToken":
 		return app.setNodeTokenCheckTx(param, nodeID)
+	case "AddAllowedNodeSupportedFeature":
+		return app.addAllowedNodeSupportedFeatureCheckTx(param, nodeID)
+	case "RemoveAllowedNodeSupportedFeature":
+		return app.removeAllowedNodeSupportedFeatureCheckTx(param, nodeID)
 
 	case "AddNamespace":
 		return app.addNamespaceCheckTx(param, nodeID)